+++ packages/neozip-blockchain/src/constants/index.ts
@@ -1,5 +1,6 @@
 /**
- * Centralized constants (metadata filenames, etc.)
+ * Centralized constants (metadata filenames, Token Service servers, etc.)
  */
 
 export * from './metadata';
+export * from './servers';